+++ assets/js/web-components/prpl-gauge.js
@@ -18,14 +18,15 @@ customElements.define(
 				max: 100,
 				value: 0,
 				maxDeg: '180deg',
-				background: 'var(--prpl-background-orange)',
-				color: 'var(--prpl-color-accent-orange)',
+				background: 'var(--prpl-background-monthly)',
+				color: 'var(--prpl-color-monthly)',
 				start: '270deg',
 				cutout: '57%',
 				contentFontSize: 'var(--prpl-font-size-6xl)',
 				contentPadding:
 					'var(--prpl-padding) var(--prpl-padding) calc(var(--prpl-padding) * 2) var(--prpl-padding)',
 				marginBottom: 'var(--prpl-padding)',
+				brandingId: 0,
 			},
 			content = ''
 		) {
@@ -60,18 +61,20 @@ customElements.define(
 				this.getAttribute( 'contentPadding' ) || props.contentPadding;
 			props.marginBottom =
 				this.getAttribute( 'marginBottom' ) || props.marginBottom;
+			props.brandingId =
+				this.getAttribute( 'branding-id' ) || props.brandingId;
 
 			this.innerHTML = `
 			<div style="padding: ${ props.contentPadding };
 			background: ${ props.background }; border-radius:var(--prpl-border-radius-big); aspect-ratio: 2 / 1; overflow: hidden; position: relative; margin-bottom: ${ props.marginBottom };">
-				<div style="width: 100%; aspect-ratio: 1 / 1; border-radius: 100%; position: relative; background: radial-gradient(${ props.background } 0 ${ props.cutout }, transparent ${ props.cutout } 100%), conic-gradient(from ${ props.start }, ${ props.color } calc(${ props.maxDeg } * ${ props.value }), var(--prpl-color-gray-1) calc(${ props.maxDeg } * ${ props.value }) ${ props.maxDeg }, transparent ${ props.maxDeg }); text-align: center;">
-					<span style="font-size: var(--prpl-font-size-small); position: absolute; top: 50%; color: var(--prpl-color-gray-5); width: 10%; text-align: center; left:0;">0</span>
-						<span style="font-size: ${ props.contentFontSize }; ${ contentSpecificStyles } display: block; font-weight: 600; text-align: center; position: absolute; color: var(--prpl-color-gray-5); width: 100%; line-height: 1.2;">
+				<div style="width: 100%; aspect-ratio: 1 / 1; border-radius: 100%; position: relative; background: radial-gradient(${ props.background } 0 ${ props.cutout }, transparent ${ props.cutout } 100%), conic-gradient(from ${ props.start }, ${ props.color } calc(${ props.maxDeg } * ${ props.value }), var(--prpl-color-gauge-remain) calc(${ props.maxDeg } * ${ props.value }) ${ props.maxDeg }, transparent ${ props.maxDeg }); text-align: center;">
+					<span style="font-size: var(--prpl-font-size-small); position: absolute; top: 50%; color: var(--prpl-color-text); width: 10%; text-align: center; left:0;">0</span>
+						<span style="font-size: ${ props.contentFontSize }; ${ contentSpecificStyles } display: block; font-weight: 600; text-align: center; position: absolute; color: var(--prpl-color-text); width: 100%; line-height: 1.2;">
 							<span style="display:inline-block;width: 50%;">
 								${ content }
 							</span>
 						</span>
-					<span style="font-size: var(--prpl-font-size-small); position: absolute; top: 50%; color: var(--prpl-color-gray-5); width: 10%; text-align: center; right:0;">${ props.max }</span>
+					<span style="font-size: var(--prpl-font-size-small); position: absolute; top: 50%; color: var(--prpl-color-text); width: 10%; text-align: center; right:0;">${ props.max }</span>
 				</div>
 			</div>
 			`;
@@ -130,8 +133,9 @@ const prplUpdateRaviGauge = ( pointsDiff ) => {
 			contentPadding:
 				'var(--prpl-padding) var(--prpl-padding) calc(var(--prpl-padding) * 2) var(--prpl-padding)',
 			marginBottom: 'var(--prpl-padding)',
+			brandingId: gaugeProps.brandingId,
 		},
-		`<prpl-badge complete="true" badge-id="${ gaugeProps.badgeId }" badge-name="${ gaugeProps.badgeName }"></prpl-badge>`
+		`<prpl-badge complete="true" badge-id="${ gaugeProps.badgeId }" badge-name="${ gaugeProps.badgeName }" branding-id="${ gaugeProps.brandingId }"></prpl-badge>`
 	);
 	gauge.id = gaugeProps.id;
 	gauge.setAttribute( 'background', gaugeProps.background );
@@ -139,6 +143,7 @@ const prplUpdateRaviGauge = ( pointsDiff ) => {
 	gauge.setAttribute( 'data-max', gaugeProps.max );
 	gauge.setAttribute( 'data-value', newValue );
 	gauge.setAttribute( 'data-badge-id', gaugeProps.badgeId );
+	gauge.setAttribute( 'data-branding-id', gaugeProps.brandingId );
 
 	// Replace the old gauge with the new one.
 	const oldGauge = document.getElementById( gaugeProps.id );