+++ classes/suggested-tasks/providers/class-tasks.php
@@ -284,7 +284,7 @@ abstract class Tasks implements Tasks_Interface {
 	 * @return string
 	 */
 	public function get_external_link_url() {
-		return static::EXTERNAL_LINK_URL;
+		return \progress_planner()->get_ui__branding()->get_url( static::EXTERNAL_LINK_URL );
 	}
 
 	/**
@@ -678,12 +678,7 @@ abstract class Tasks implements Tasks_Interface {
 		}
 
 		// Order actions by priority.
-		\usort(
-			$actions,
-			function ( $a, $b ) {
-				return $a['priority'] - $b['priority'];
-			}
-		);
+		\usort( $actions, fn( $a, $b ) => $a['priority'] - $b['priority'] );
 
 		$return_actions = [];
 		foreach ( $actions as $action ) {
@@ -703,5 +698,27 @@ abstract class Tasks implements Tasks_Interface {
 	 */
 	public function add_task_actions( $data = [], $actions = [] ) {
 		return $actions;
+	}
+
+	/**
+	 * Check if the task has activity.
+	 *
+	 * @param string $task_id The task ID.
+	 *
+	 * @return bool
+	 */
+	public function task_has_activity( $task_id = '' ) {
+		if ( empty( $task_id ) ) {
+			$task_id = $this->get_task_id();
+		}
+
+		$activity = \progress_planner()->get_activities__query()->query_activities(
+			[
+				'category' => 'suggested_task',
+				'data_id'  => $task_id,
+			]
+		);
+
+		return ! empty( $activity );
 	}
 }