+++ classes/utils/class-system-status.php
@@ -72,27 +72,20 @@ class System_Status {
 
 		$scores = \progress_planner()->get_ui__chart()->get_chart_data(
 			[
-				'items_callback' => function ( $start_date, $end_date ) {
-					return \progress_planner()->get_activities__query()->query_activities(
-						[
-							'start_date' => $start_date,
-							'end_date'   => $end_date,
-						]
-					);
-				},
+				'items_callback' => fn( $start_date, $end_date ) => \progress_planner()->get_activities__query()->query_activities(
+					[
+						'start_date' => $start_date,
+						'end_date'   => $end_date,
+					]
+				),
 				'dates_params'   => [
 					'start_date' => \DateTime::createFromFormat( 'Y-m-d', \gmdate( 'Y-m-01' ) )->modify( '-6 months' ),
 					'end_date'   => new \DateTime(),
 					'frequency'  => 'monthly',
 					'format'     => 'M',
 				],
-				'count_callback' => function ( $activities, $date ) {
-					$score = 0;
-					foreach ( $activities as $activity ) {
-						$score += $activity->get_points( $date );
-					}
-					return $score * 100 / Base::SCORE_TARGET;
-				},
+				'count_callback' => fn( $activities, $date ) =>
+					\array_sum( \array_map( fn( $activity ) => $activity->get_points( $date ), $activities ) ) * 100 / Base::SCORE_TARGET,
 				'normalized'     => true,
 				'max'            => 100,
 			]
@@ -139,6 +132,8 @@ class System_Status {
 				'version' => $plugin_data['Version'] ?? 'N/A', // @phpstan-ignore-line nullCoalesce.offset
 			];
 		}
+
+		$data['branding_id'] = (int) \progress_planner()->get_ui__branding()->get_branding_id();
 
 		return $data;
 	}