+++ views/page-widgets/content-activity.php
@@ -14,11 +14,11 @@ $prpl_widget = \progress_planner()->get_admin__widgets__content_activity();
 $prpl_activity_types = [
 	'publish' => [
 		'label' => \__( 'published', 'progress-planner' ),
-		'color' => 'var(--prpl-color-accent-orange)',
+		'color' => 'var(--prpl-color-monthly)',
 	],
 	'update'  => [
 		'label' => \__( 'updated', 'progress-planner' ),
-		'color' => 'var(--prpl-color-accent-green)',
+		'color' => 'var(--prpl-graph-color-3)',
 	],
 	'delete'  => [
 		'label' => \__( 'deleted', 'progress-planner' ),
@@ -35,8 +35,8 @@ $prpl_chart_data    = [];
 $prpl_chart_options = [
 	'dataArgs'     => [],
 	'chartId'      => 'prpl-chart-content-activity',
-	'axisColor'    => 'var(--prpl-color-gray-2)',
-	'rulersColor'  => 'var(--prpl-color-gray-2)',
+	'axisColor'    => 'var(--prpl-color-border)',
+	'rulersColor'  => 'var(--prpl-color-border)',
 	'filtersLabel' => '<strong>' . \__( 'show:', 'progress-planner' ) . '</strong>',
 ];
 foreach ( $prpl_activity_types as $prpl_activity_type => $prpl_activity_data ) {
@@ -72,13 +72,7 @@ foreach ( \array_keys( $prpl_activity_types ) as $prpl_activity_type ) {
 	if ( $prpl_activities ) {
 		if ( 'delete' !== $prpl_activity_type ) {
 			// Filter the activities to only include the tracked post types.
-			$prpl_activities = \array_filter(
-				$prpl_activities,
-				function ( $activity ) use ( $prpl_tracked_post_types ) {
-
-					return \in_array( \get_post_type( $activity->data_id ), $prpl_tracked_post_types, true );
-				}
-			);
+			$prpl_activities = \array_filter( $prpl_activities, fn( $activity ) => \in_array( \get_post_type( $activity->data_id ), $prpl_tracked_post_types, true ) );
 		}
 
 		// Update the count.
@@ -90,16 +84,19 @@ foreach ( \array_keys( $prpl_activity_types ) as $prpl_activity_type ) {
 ?>
 
 <h2 class="prpl-widget-title">
-	<?php \esc_html_e( 'Content activity', 'progress-planner' ); ?>
+	<?php
+	echo \progress_planner()->get_ui__branding()->get_widget_title( // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+		'content-activity',
+		\esc_html__( 'Content activity', 'progress-planner' )
+	);
+	?>
 </h2>
-<p>
-	<?php \esc_html_e( 'Here are the updates you made to your content last week. Whether you published something new, updated an existing post, or removed outdated content, it all helps you stay on top of your site!', 'progress-planner' ); ?>
-</p>
+<p><?php \esc_html_e( 'Here are the updates you made to your content last week. Whether you published something new, updated an existing post, or removed outdated content, it all helps you stay on top of your site!', 'progress-planner' ); ?></p>
 
 <prpl-big-counter
 	number="<?php echo \esc_html( \number_format_i18n( (int) $prpl_activities_count['all'] ) ); ?>"
 	content="<?php \esc_attr_e( 'pieces of content managed', 'progress-planner' ); ?>"
-	background-color="var(--prpl-background-purple)"
+	background-color="var(--background-content)"
 ></prpl-big-counter>
 
 <div class="prpl-graph-wrapper">