+++ views/page-widgets/parts/monthly-badges.php
@@ -41,12 +41,23 @@ foreach ( \progress_planner()->get_admin__widgets__monthly_badges()->get_previou
 		<?php
 		if ( 'popover' !== $prpl_location && 3 < \count( $prpl_badges ) ) {
 			// If we have more than 3 badges, we need to show the current month badge and the previous 2 months badges.
-			$prpl_current_month_badge_id = Monthly::get_badge_id_from_date( new \DateTime() );
-			$prpl_temp_badges            = [];
+			$prpl_current_month_badge_id    = Monthly::get_badge_id_from_date( new \DateTime() );
+			$prpl_temp_badges               = [];
+			$prpl_current_month_badge_index = 0;
 			foreach ( $prpl_badges as $prpl_badge ) {
 				$prpl_temp_badges[] = $prpl_badge;
 				if ( $prpl_current_month_badge_id === $prpl_badge->get_id() ) {
 					break;
+				}
+				++$prpl_current_month_badge_index;
+			}
+			if ( 3 > \count( $prpl_temp_badges ) ) {
+				$prpl_temp_badges = [ $prpl_badges[ $prpl_current_month_badge_index ] ];
+				if ( isset( $prpl_badges[ $prpl_current_month_badge_index + 1 ] ) ) {
+					$prpl_temp_badges[] = $prpl_badges[ $prpl_current_month_badge_index + 1 ];
+				}
+				if ( isset( $prpl_badges[ $prpl_current_month_badge_index + 2 ] ) ) {
+					$prpl_temp_badges[] = $prpl_badges[ $prpl_current_month_badge_index + 2 ];
 				}
 			}
 			$prpl_badges = \array_slice( $prpl_temp_badges, -3 ); // We show only 3 badges in page widget.
@@ -65,6 +76,7 @@ foreach ( \progress_planner()->get_admin__widgets__monthly_badges()->get_previou
 							complete="<?php echo 100 === (int) $prpl_badge->progress_callback()['progress'] ? 'true' : 'false'; ?>"
 							badge-id="<?php echo \esc_attr( $prpl_badge->get_id() ); ?>"
 							badge-name="<?php echo \esc_attr( $prpl_badge->get_name() ); ?>"
+							branding-id="<?php echo (int) \progress_planner()->get_ui__branding()->get_branding_id(); ?>"
 						></prpl-badge>
 						<p><?php echo \esc_html( $prpl_badge->get_name() ); ?></p>
 					</span>